feat: id로 카트 상품 삭제

+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -29,9 +29,9 @@ class CartViewModel(
         }
     }
 
-    fun deleteCartProduct(id: Int) {
+    fun deleteCartProduct(id: Long) {
         viewModelScope.launch {
-            cartRepository.deleteCartProduct(id)
+            cartRepository.deleteCartProduct(id.toInt())
             _onCartProductDeleted.value = true
         }
     }
